Multimodal-RAG/requirements.txt
# Core Framework
streamlit>=1.30.0

# RAG Orchestration
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.1.0

# Vector Database
chromadb>=0.4.22

# Audio Transcription
openai-whisper>=20231117

# Vision & Embeddings
open-clip-torch>=2.24.0
sentence-transformers>=2.3.0

# Video Processing
yt-dlp>=2024.1.0
opencv-python>=4.9.0
scenedetect[opencv]>=0.6.3
imagehash>=4.3.1
moviepy==1.0.3

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# Image Processing
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
